refacto: création d'un composant CreateNewTeam qui englobe l'ajout d'une nouvelle équipe
